Mulighet for å filtrere "Endre komponenter" etter pris, og mulighet for å vise alle komponentene. Også gjort det mulig for brukeren å endre opplysninger om seg selv.

+++ src/main/java/data/KomponentData.java
@@ -42,7 +42,7 @@ public class KomponentData {
         tv.setItems(utvalgteLagrinsenheter);
     }
 
-    public void sorterEtterPris(TableView tv, int prisFra, int prisTil, String enhet){
+    public void sorterEtterPris(TableView tv, double prisFra, double prisTil, String enhet){
         ObservableList<Komponent> returListe = FXCollections.observableArrayList();
         for(Komponent enKomponent : alleKomponenter){
             //Sjekker om prisen er innenfor grensene
